Changed Types and Added attributes to some fields of Classes of Model Class Libraries

+++ StockMarket.Models/StockExchange.cs
@@ -7,7 +7,7 @@ namespace StockMarket.Models
     class StockExchange
     {
         public long Id { get; set; }
-        public int StockExchangeValue { get; set; }
+        public decimal StockExchangeValue { get; set; }
         public string Brief { get; set; }
         public string ContactAddress { get; set; }
         public string Remarks { get; set; }
